llm-runtime/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1
faiss-cpu>=1.7.4
pypdf>=3.17.0
httpx>=0.25.0
sentence-transformers>=2.2.2
python-multipart
google-generativeai>=0.3.0
python-pptx>=0.6.21
matplotlib>=3.7.0
Pillow>=10.0.0


pdfplumber>=0.10.0
